Refactor the testing functions to a TestDataUtils class
